# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Foodgram'
  version: ''
paths:
...
  /api/recipes/{id}/favorite/:
    post:
      operationId: Добавить рецепт в избранное
      description: 'Доступно только авторизованному пользователю.'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeMinified'
          description: 'Рецепт успешно добавлен в избранное'
        '400':
          description: 'Ошибка добавления в избранное (Например, когда рецепт уже есть в избранном)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'

      tags:
        - Избранное
    delete:
      operationId: Удалить рецепт из избранного
      description: 'Доступно только авторизованным пользователям'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален из избранного'
        '400':
          description: 'Ошибка удаления из избранного (Например, когда рецепта там не было)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Избранное
...
  /api/ingredients/:
    get:
      operationId: Список ингредиентов
      description: 'Список ингредиентов с возможностью поиска по имени.'
      parameters:
        - name: name
          required: false
          in: query
          description: Поиск по частичному вхождению в начале названия ингредиента.
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
...
components:
  schemas:
...
    RecipeMinified:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        name:
          type: string
          maxLength: 200
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://foodgram.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          example: 'Капуста'
        measurement_unit:
          type: string
          maxLength: 200
          example: 'кг'
      required:
        - name
        - measurement_unit
...
    SelfMadeError:
      description: Ошибка
      type: object
      properties:
        errors:
          description: 'Описание ошибки'
          type: string
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
...
  responses:
...
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
       Все запросы от имени пользователя должны выполняться 
       с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token